requirements.txt
streamlit
llama-index
chromadb
spacy
rake-nltk
scikit-learn
PyMuPDF
python-dotenv
llama-index-embeddings-bedrock
llama-index-llms-bedrock
llama-index-vector-stores-chroma
langdetect
py_vncorenlp
boto3
pillow>=12.0.0
